Add files via upload
The two test/sample code files contain test master/slave code for the new axis. The ModulaAxisV2 file integrates both master and slave functions into one script. Implemented are PWM, I2C communication, the ability for the master to send values to the slave. This needs to be integrated with original Arduino file for Unity to be able to send data through the Serial port to the system.